Move Blog, and Category models into Models folder
